requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
pytest==7.4.3
openai==1.12.0
python-mcp-sdk>=0.6.0
